RedisCluster support
To use clustered Redis, define a cluster in the INI file and set
session.save_path to the cluster name. Note that this differs from the
Redis module's session.save_path; this is a simpler option tailored to
our current needs.

FAILOVER_DISTRIBUTE is set to distribute commands between masters and
slaves.